fix: close connection on error during encoding or flushing Responses
Modify WriteResponse to call cc.Close() immediately if an error occurs
while encoding the response header, encoding the body, or flushing the
buffered writer. This prevents further use of a codec that encountered
an error, ensuring that subsequent operations return io.EOF. This is
*not* needed for WriteRequest as net/rpc propagates that error to the
caller, and a comment has been added to that method explaining the
difference.